use oapconf to avoid override sqlconf
