+++ src/main/scala/org/apache/spark/sql/execution/datasources/oap/OapFileFormat.scala
@@ -28,7 +28,6 @@ import org.apache.hadoop.mapreduce.{Job, TaskAttemptContext}
 import org.apache.hadoop.mapreduce.lib.output.FileOutputFormat
 import org.apache.parquet.hadoop.util.SerializationUtil
 
-import org.apache.spark.{SparkContext, TaskContext}
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.{Row, SparkSession}
 import org.apache.spark.sql.catalyst.InternalRow
@@ -39,8 +38,8 @@ import org.apache.spark.sql.execution.datasources.oap.filecache.DataFileHandleCa
 import org.apache.spark.sql.execution.datasources.oap.index.{IndexContext, ScannerBuilder}
 import org.apache.spark.sql.execution.datasources.oap.io._
 import org.apache.spark.sql.execution.datasources.oap.utils.OapUtils
-import org.apache.spark.sql.execution.metric.{SQLMetric, SQLMetrics}
-import org.apache.spark.sql.internal.SQLConf
+import org.apache.spark.sql.execution.metric.SQLMetric
+import org.apache.spark.sql.internal.oap.OapConf
 import org.apache.spark.sql.sources._
 import org.apache.spark.sql.types.{StructField, StructType}
 import org.apache.spark.util.SerializableConfiguration
@@ -107,14 +106,14 @@ private[sql] class OapFileFormat extends FileFormat
     // TODO: Should we have our own config util instead of SqlConf?
     // First use table option, if not, use SqlConf, else, use default value.
     conf.set(OapFileFormat.COMPRESSION, options.getOrElse("compression",
-        sparkSession.conf.get(SQLConf.OAP_COMPRESSION.key,
+        sparkSession.conf.get(OapConf.OAP_COMPRESSION.key,
           OapFileFormat.DEFAULT_COMPRESSION)))
 
     conf.set(OapFileFormat.ROW_GROUP_SIZE, options.getOrElse("rowgroup",
-      sparkSession.conf.get(SQLConf.OAP_ROW_GROUP_SIZE.key,
+      sparkSession.conf.get(OapConf.OAP_ROW_GROUP_SIZE.key,
       OapFileFormat.DEFAULT_ROW_GROUP_SIZE)))
 
-    new OapOutputWriterFactory(sparkSession.sqlContext.conf,
+    new OapOutputWriterFactory(
       dataSchema,
       job,
       options)
@@ -257,7 +256,7 @@ private[sql] class OapFileFormat extends FileFormat
             supportFilters.foreach(filter => logDebug("\t" + filter.toString))
             // get index options such as limit, order, etc.
             val indexOptions = options.filterKeys(OapFileFormat.oapOptimizationKeySeq.contains(_))
-            val maxChooseSize = sparkSession.conf.get(SQLConf.OAP_INDEXER_CHOICE_MAX_SIZE)
+            val maxChooseSize = sparkSession.conf.get(OapConf.OAP_INDEXER_CHOICE_MAX_SIZE)
             ScannerBuilder.build(supportFilters, ic, indexOptions, maxChooseSize)
           }
         }
@@ -273,10 +272,10 @@ private[sql] class OapFileFormat extends FileFormat
 
         val requiredIds = requiredSchema.map(dataSchema.fields.indexOf(_)).toArray
 
-        hadoopConf.setDouble(SQLConf.OAP_FULL_SCAN_THRESHOLD.key,
-          sparkSession.conf.get(SQLConf.OAP_FULL_SCAN_THRESHOLD))
-        hadoopConf.setBoolean(SQLConf.OAP_ENABLE_OINDEX.key,
-          sparkSession.conf.get(SQLConf.OAP_ENABLE_OINDEX))
+        hadoopConf.setDouble(OapConf.OAP_FULL_SCAN_THRESHOLD.key,
+          sparkSession.conf.get(OapConf.OAP_FULL_SCAN_THRESHOLD))
+        hadoopConf.setBoolean(OapConf.OAP_ENABLE_OINDEX.key,
+          sparkSession.conf.get(OapConf.OAP_ENABLE_OINDEX))
         val broadcastedHadoopConf =
           sparkSession.sparkContext.broadcast(new SerializableConfiguration(hadoopConf))
 
@@ -332,7 +331,7 @@ private[sql] class OapFileFormat extends FileFormat
   def hasAvailableIndex(
       expressions: Seq[Expression],
       requiredTypes: Seq[IndexType] = Nil): Boolean = {
-    if (expressions.nonEmpty && sparkSession.conf.get(SQLConf.OAP_ENABLE_OINDEX)) {
+    if (expressions.nonEmpty && sparkSession.conf.get(OapConf.OAP_ENABLE_OINDEX)) {
       meta match {
         case Some(m) if requiredTypes.isEmpty =>
           expressions.exists(m.isSupportedByIndex(_, None))
@@ -355,13 +354,11 @@ private[oap] object INDEX_STAT extends Enumeration {
 
 /**
  * Oap Output Writer Factory
- * @param sqlConf
  * @param dataSchema
  * @param job
  * @param options
  */
 private[oap] class OapOutputWriterFactory(
-    sqlConf: SQLConf,
     dataSchema: StructType,
     @transient protected val job: Job,
     options: Map[String, String]) extends OutputWriterFactory {
@@ -499,9 +496,9 @@ private[sql] object OapFileFormat {
   val PARQUET_DATA_FILE_CLASSNAME = classOf[ParquetDataFile].getCanonicalName
 
   val COMPRESSION = "oap.compression"
-  val DEFAULT_COMPRESSION = SQLConf.OAP_COMPRESSION.defaultValueString
+  val DEFAULT_COMPRESSION = OapConf.OAP_COMPRESSION.defaultValueString
   val ROW_GROUP_SIZE = "oap.rowgroup.size"
-  val DEFAULT_ROW_GROUP_SIZE = SQLConf.OAP_ROW_GROUP_SIZE.defaultValueString
+  val DEFAULT_ROW_GROUP_SIZE = OapConf.OAP_ROW_GROUP_SIZE.defaultValueString
 
   def serializeDataSourceMeta(conf: Configuration, meta: Option[DataSourceMeta]): Unit = {
     SerializationUtil.writeObjectToConfAsBase64(OAP_DATA_SOURCE_META, meta, conf)
