+++ src/test/scala/org/apache/spark/sql/execution/datasources/oap/FilterSuite.scala
@@ -25,7 +25,7 @@ import org.scalatest.BeforeAndAfterEach
 
 import org.apache.spark.sql.{QueryTest, Row}
 import org.apache.spark.sql.catalyst.util.DateTimeUtils
-import org.apache.spark.sql.internal.SQLConf
+import org.apache.spark.sql.internal.oap.OapConf
 import org.apache.spark.sql.test.oap.SharedOapContext
 import org.apache.spark.sql.types.{IntegerType, StringType, StructField, StructType}
 import org.apache.spark.util.Utils
@@ -85,9 +85,9 @@ class FilterSuite extends QueryTest with SharedOapContext with BeforeAndAfterEac
   }
 
   test("test oap row group size change") {
-    val previousRowGroupSize = sqlConf.getConfString(SQLConf.OAP_ROW_GROUP_SIZE.key)
+    val previousRowGroupSize = sqlConf.getConfString(OapConf.OAP_ROW_GROUP_SIZE.key)
     // change default row group size
-    sqlConf.setConfString(SQLConf.OAP_ROW_GROUP_SIZE.key, "1025")
+    sqlConf.setConfString(OapConf.OAP_ROW_GROUP_SIZE.key, "1025")
     val data: Seq[(Int, String)] = (1 to 3000).map { i => (i, s"this is test $i") }
     data.toDF("key", "value").createOrReplaceTempView("t")
     checkAnswer(sql("SELECT * FROM oap_test"), Seq.empty[Row])
@@ -95,10 +95,10 @@ class FilterSuite extends QueryTest with SharedOapContext with BeforeAndAfterEac
     checkAnswer(sql("SELECT * FROM oap_test"), data.map { row => Row(row._1, row._2) })
     // set back to default value
     if (previousRowGroupSize == null) {
-      sqlConf.setConfString(SQLConf.OAP_ROW_GROUP_SIZE.key,
-        SQLConf.OAP_ROW_GROUP_SIZE.defaultValueString)
+      sqlConf.setConfString(OapConf.OAP_ROW_GROUP_SIZE.key,
+        OapConf.OAP_ROW_GROUP_SIZE.defaultValueString)
     } else {
-      sqlConf.setConfString(SQLConf.OAP_ROW_GROUP_SIZE.key,
+      sqlConf.setConfString(OapConf.OAP_ROW_GROUP_SIZE.key,
         previousRowGroupSize)
     }
   }
