+++ src/test/scala/org/apache/spark/sql/execution/datasources/oap/statistics/BloomFilterStatisticsSuite.scala
@@ -23,7 +23,7 @@ import org.apache.hadoop.conf.Configuration
 
 import org.apache.spark.sql.catalyst.expressions.{BoundReference, UnsafeProjection}
 import org.apache.spark.sql.execution.datasources.oap.index.{BloomFilter, IndexUtils}
-import org.apache.spark.sql.internal.SQLConf
+import org.apache.spark.sql.internal.oap.OapConf
 import org.apache.spark.sql.types.StructType
 
 
@@ -54,8 +54,8 @@ class BloomFilterStatisticsSuite extends StatisticsTest {
   test("write function test") {
     val keys = Random.shuffle(1 to 300).map(i => rowGen(i)).toArray
 
-    val maxBits = SQLConf.OAP_BLOOMFILTER_MAXBITS.defaultValue.get
-    val numOfHashFunc = SQLConf.OAP_BLOOMFILTER_NUMHASHFUNC.defaultValue.get
+    val maxBits = OapConf.OAP_BLOOMFILTER_MAXBITS.defaultValue.get
+    val numOfHashFunc = OapConf.OAP_BLOOMFILTER_NUMHASHFUNC.defaultValue.get
     val bfIndex = new BloomFilter(maxBits, numOfHashFunc)()
     val boundReference = schema.zipWithIndex.map(x =>
       BoundReference(x._2, x._1.dataType, nullable = true))
@@ -93,8 +93,8 @@ class BloomFilterStatisticsSuite extends StatisticsTest {
   test("read function test") {
     val keys = Random.shuffle(1 to 300).map(i => rowGen(i)).toArray
 
-    val maxBits = SQLConf.OAP_BLOOMFILTER_MAXBITS.defaultValue.get
-    val numOfHashFunc = SQLConf.OAP_BLOOMFILTER_NUMHASHFUNC.defaultValue.get
+    val maxBits = OapConf.OAP_BLOOMFILTER_MAXBITS.defaultValue.get
+    val numOfHashFunc = OapConf.OAP_BLOOMFILTER_NUMHASHFUNC.defaultValue.get
     val bfIndex = new BloomFilter(maxBits, numOfHashFunc)()
     val boundReference = schema.zipWithIndex.map(x =>
       BoundReference(x._2, x._1.dataType, nullable = true))
@@ -122,8 +122,8 @@ class BloomFilterStatisticsSuite extends StatisticsTest {
   test("read AND write test") {
     val keys = Random.shuffle(1 to 300).map(i => rowGen(i)).toArray
 
-    val maxBits = SQLConf.OAP_BLOOMFILTER_MAXBITS.defaultValue.get
-    val numOfHashFunc = SQLConf.OAP_BLOOMFILTER_NUMHASHFUNC.defaultValue.get
+    val maxBits = OapConf.OAP_BLOOMFILTER_MAXBITS.defaultValue.get
+    val numOfHashFunc = OapConf.OAP_BLOOMFILTER_NUMHASHFUNC.defaultValue.get
     val bfIndex = new BloomFilter(maxBits, numOfHashFunc)()
     val boundReference = schema.zipWithIndex.map(x =>
       BoundReference(x._2, x._1.dataType, nullable = true))
@@ -155,8 +155,8 @@ class BloomFilterStatisticsSuite extends StatisticsTest {
   test("test analyze function") {
     val keys = Random.shuffle(1 to 300).map(i => rowGen(i)).toArray
 
-    val maxBits = SQLConf.OAP_BLOOMFILTER_MAXBITS.defaultValue.get
-    val numOfHashFunc = SQLConf.OAP_BLOOMFILTER_NUMHASHFUNC.defaultValue.get
+    val maxBits = OapConf.OAP_BLOOMFILTER_MAXBITS.defaultValue.get
+    val numOfHashFunc = OapConf.OAP_BLOOMFILTER_NUMHASHFUNC.defaultValue.get
     val bfIndex = new BloomFilter(maxBits, numOfHashFunc)()
     val boundReference = schema.zipWithIndex.map(x =>
       BoundReference(x._2, x._1.dataType, nullable = true))
